--- src/components/Cart/Cart.css
@@ -1,3 +1,0 @@
-.cartycart {
-	background-color: var(--colorPrimary);
-}
