+++ src/components/Cart/CartComponent.css
@@ -1,0 +1,4 @@
+.cartycart {
+	background-color: var(--colorLight);
+	color: var(--colorDark);
+}
